Fixed Wifi Connection for the robot
Set up the wifi connection to work on my pc (for testing). Be sure to change the IP address to whatever your computer's is (locally ofc). On windows you can do "ipconfig" in the cmd and copy the first IPV4 address you find. Be sure to change it back before merging.